# openapi.yaml
openapi: 3.0.3
info:
  title: Hệ thống Hoán đổi Khuôn mặt và Giọng nói
  description: |
    Hệ thống API tích hợp nhiều tính năng xử lý AI:

    - 🎭 Hoán đổi khuôn mặt giữa hai ảnh
    - 🗣️ Chuyển văn bản thành giọng nói
    - 🖼️ Tải ảnh và tệp âm thanh
    - 🎬 Tạo video hoạt hình từ ảnh và âm thanh
    - 🧠 Chuyển đổi và quản lý mô hình AI động

    Công nghệ sử dụng: vietTTS, Wav2Lip, InsightFace, OpenAI...
  version: 1.0.0
  contact:
    name: Nhóm phát triển AI
    email: support@example.com
servers:
  - url: http://localhost:55021
    description: Local development server
...
paths:
...
  /face_swap/:
    post:
      summary: Hoán đổi khuôn mặt giữa hai hình ảnh
      description: |
        Nhận hai hình ảnh đầu vào: ảnh nguồn (chứa khuôn mặt cần chuyển) và ảnh đích (nơi khuôn mặt sẽ được áp dụng). API sử dụng mô hình AI (mặc định là `inswapper_128.onnx`) để thực hiện hoán đổi khuôn mặt, tạo ra một hình ảnh mới với khuôn mặt từ ảnh nguồn được tích hợp tự nhiên vào ảnh đích.

        **Cách sử dụng**:
        - Tải lên hai tệp hình ảnh (định dạng JPG, PNG, v.v.) thông qua form-data.
        - API sẽ xử lý và lưu kết quả vào thư mục `static` với tên `output.jpg`.
        - Kiểm tra phản hồi JSON để xác nhận quá trình hoàn tất và lấy thông tin về mô hình đã sử dụng.

        **Lưu ý**:
        - Đảm bảo cả hai ảnh đều chứa ít nhất một khuôn mặt rõ ràng để đạt kết quả tốt nhất.
        - Kích thước và chất lượng ảnh có thể ảnh hưởng đến kết quả; nên sử dụng ảnh có độ phân giải cao.
        - Có thể thay đổi mô hình AI thông qua endpoint `/change_model/` để tùy chỉnh hiệu suất hoặc chất lượng.
        - Xử lý có thể mất vài giây tùy thuộc vào tài nguyên máy tính và kích thước ảnh.
      operationId: face_swap
      tags:
        - Xử lý ảnh
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                source:
                  type: string
                  format: binary
                  description: Ảnh chứa khuôn mặt cần chuyển
                target:
                  type: string
                  format: binary
                  description: Ảnh chứa khuôn mặt đích
              required:
                - source
                - target
      responses:
        '200':
          description: Hoán đổi khuôn mặt thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Face swap completed!
                  model_used:
                    type: string
                    example: /path/to/model/inswapper_128.onnx
        '500':
          description: Lỗi khi thực hiện hoán đổi khuôn mặt
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /text_to_speech/:
    post:
      summary: Chuyển văn bản thành giọng nói
      description: |
        Sử dụng công cụ vietTTS để chuyển đổi văn bản đầu vào thành tệp âm thanh WAV với giọng nói tiếng Việt được chọn. API hỗ trợ nhiều giọng nói khác nhau (được đánh số từ 1 đến 24), cho phép tạo ra âm thanh tự nhiên và phù hợp với các ứng dụng như lồng tiếng, trợ lý ảo, hoặc nội dung giáo dục.

        **Cách sử dụng**:
        - Cung cấp văn bản cần chuyển đổi và mã số giọng nói thông qua form-data.
        - API sẽ tạo tệp âm thanh và lưu vào thư mục `static` với tên `output.wav`.
        - Phản hồi JSON chứa đường dẫn đến tệp âm thanh đầu ra.

        **Lưu ý**:
        - Văn bản nên ngắn gọn và không chứa ký tự đặc biệt phức tạp để đảm bảo chất lượng âm thanh.
        - Kiểm tra danh sách giọng nói hợp lệ (1-24) trong tài liệu của vietTTS trước khi sử dụng.
        - Tệp WAV đầu ra có thể được sử dụng trực tiếp hoặc chuyển đổi sang định dạng khác (như MP3) nếu cần.
        - Xử lý nhanh nhưng có thể bị ảnh hưởng nếu hệ thống vietTTS chưa được cài đặt hoặc cấu hình đúng.
      operationId: text_to_speech
      tags:
        - Chuyển đổi giọng nói
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: Nội dung văn bản cần chuyển thành giọng nói
                voice:
                  type: string
                  description: Tên giọng nói sẽ sử dụng (từ 1 đến 24)
              required:
                - text
                - voice
      responses:
        '200':
          description: Chuyển đổi văn bản thành giọng nói thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  output_audio:
                    type: string
                    example: /static/output.wav
        '500':
          description: Lỗi khi thực hiện chuyển đổi
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /animate_image/:
    post:
      summary: Tạo video hoạt hình từ ảnh và âm thanh
      description: |
        Tạo một video hoạt hình bằng cách kết hợp một hình ảnh tĩnh (`output.jpg`) và một tệp âm thanh (`output.wav`) sử dụng công cụ Wav2Lip. API đồng bộ chuyển động khuôn mặt trong ảnh với âm thanh, tạo hiệu ứng như nhân vật trong ảnh đang nói, phù hợp cho các video quảng cáo, giải trí, hoặc giáo dục.

        **Cách sử dụng**:
        - Đảm bảo tệp `output.jpg` và `output.wav` đã tồn tại trong thư mục `static` (có thể tải lên Alt+Ctrl/Cmd+Enter để tải lên qua các endpoint `/upload_image` và `/upload_voice`).
        - Gửi yêu cầu POST đến endpoint này để tạo video.
        - Video đầu ra được lưu vào thư mục `static/Videos` với tên `output_web.mp4` (hoặc tên khác nếu tệp đã tồn tại).

        **Lưu ý**:
        - Ảnh cần chứa một khuôn mặt rõ ràng để Wav2Lip có thể đồng bộ chính xác.
        - Tệp âm thanh nên có chất lượng tốt và thời lượng phù hợp để video trông tự nhiên.
        - API sử dụng ffmpeg để mã hóa video đầu ra thành định dạng MP4, vì vậy cần cài đặt ffmpeg trên hệ thống.
        - Quá trình tạo video có thể mất thời gian, đặc biệt với ảnh hoặc âm thanh dài.
        - Các video cũ trong thư mục `static` sẽ bị xóa trước khi tạo video mới để tránh xung đột.
      operationId: animate_image
      tags:
        - Tạo video
      responses:
        '200':
          description: Tạo video thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  output_video:
                    type: string
                    example: /static/Videos/output_web.mp4
        '400':
          description: Thiếu tệp đầu vào (ảnh hoặc âm thanh)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Lỗi khi tạo video
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /upload_image:
    post:
      summary: Tải ảnh lên
      description: |
        Tải một tệp hình ảnh lên hệ thống và lưu vào thư mục `static` với tên `output.jpg`. Tệp này có thể được sử dụng làm đầu vào cho các quá trình như hoán đổi khuôn mặt hoặc tạo video hoạt hình.

        **Cách sử dụng**:
        - Tải lên một tệp hình ảnh (JPG, PNG, v.v.) thông qua form-data.
        - API sẽ ghi đè bất kỳ tệp `output.jpg` hiện có trong thư mục `static`.
        - Phản hồi JSON xác nhận quá trình tải lên thành công.

        **Lưu ý**:
        - Đảm bảo ảnh có định dạng hợp lệ và kích thước phù hợp để tránh lỗi xử lý.
        - Nếu đã có tệp `output.jpg`, nó sẽ bị ghi đè mà không có cảnh báo.
        - Ảnh tải lên nên chứa khuôn mặt rõ ràng nếu dự định sử dụng cho hoán đổi khuôn mặt hoặc tạo video.
      operationId: upload_image
      tags:
        - Tải tệp
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Ảnh cần tải lên
              required:
                - image
      responses:
        '200':
          description: Tải ảnh thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tải ảnh thành công
        '500':
          description: Lỗi khi tải ảnh
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /upload_voice:
    post:
      summary: Tải tệp âm thanh lên
      description: |
        Tải một tệp âm thanh lên hệ thống và lưu vào thư mục `static` với tên `output.wav`. Tệp này có thể được sử dụng làm đầu vào cho quá trình tạo video hoạt hình.

        **Cách sử dụng**:
        - Tải lên một tệp âm thanh (WAV, MP3, v.v.) thông qua form-data.
        - API sẽ ghi đè bất kỳ tệp `output.wav` hiện có trong thư mục `static`.
        - Phản hồi JSON xác nhận quá trình tải lên thành công.

        **Lưu ý**:
        - Đảm bảo âm thanh có định dạng hợp lệ và chất lượng tốt để đảm bảo đồng bộ chính xác trong video.
        - Nếu đã có tệp `output.wav`, nó sẽ bị ghi đè mà không có cảnh báo.
        - Tệp WAV là định dạng ưu tiên để đảm bảo tương thích với Wav2Lip.
      operationId: upload_voice
      tags:
        - Tải tệp
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                voice:
                  type: string
                  format: binary
                  description: Tệp âm thanh cần tải lên
              required:
                - voice
      responses:
        '200':
          description: Tải âm thanh thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tải giọng nói thành công
        '500':
          description: Lỗi khi tải âm thanh
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /models/{model_path:path}:
    get:
      summary: Phục vụ tệp mô hình
      description: |
        Truy xuất và trả về một tệp mô hình từ thư mục `models` dựa trên đường dẫn được cung cấp. API được sử dụng để tải xuống hoặc kiểm tra các mô hình đã được tải lên trước đó.

        **Cách sử dụng**:
        - Gửi yêu cầu GET với đường dẫn tương đối đến tệp mô hình (ví dụ: `models/new_model.onnx`).
        - API sẽ trả về tệp dưới dạng luồng nhị phân nếu tệp tồn tại.
        - Sử dụng đường dẫn từ phản hồi của endpoint `/change_model/` để truy cập đúng tệp.

        **Lưu ý**:
        - Đường dẫn phải chính xác và tệp phải tồn tại trong thư mục `models`.
        - API trả về tệp dưới dạng octet-stream, phù hợp để tải xuống hoặc sử dụng trực tiếp.
        - Đảm bảo quyền truy cập vào thư mục `models` để tránh lỗi.
      operationId: serve_model
      tags:
        - Quản lý mô hình
      parameters:
        - name: model_path
          in: path
          required: true
          schema:
            type: string
          description: Đường dẫn tương đối đến tệp mô hình
      responses:
        '200':
          description: Tệp mô hình được trả về thành công
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Không tìm thấy tệp mô hình
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
...